requirements.txt
absl-py==2.1.0
cloudpickle==3.1.0
Cython==0.29.37
gitdb>=4.0.11
glfw==2.7.0
gymnasium==0.28.1
gymnasium-robotics==1.2.2
imageio>=2.31.5
importlib_resources>=6.4.0
iniconfig>=2.0.0
matplotlib==3.8.0
mujoco==2.3.3
numpy==1.23.5
opencv-python==4.10.0.84
pandas==2.2.1
Pillow==10.1.0
protobuf==5.26.1
pygame==2.1.0
PyOpenGL==3.1.7
PyYAML==6.0.1
scikit-learn==1.5.1
safety-gymnasium==1.0.0
scipy==1.14.1
seaborn==0.13.2
sympy==1.13.3
tensorboard>=2.16.2
torch==2.2.2
tqdm==4.66.4
wandb==0.17.5
xmltodict==0.14.2